Fix merge problems
