+++ bookie-app/app/src/main/java/com/example/bookie/ui/book_list/BookList.kt
@@ -5,11 +5,10 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.bookie.MyApplication
 import com.example.bookie.R
 import com.example.bookie.models.BookFeed
 import com.example.bookie.repositories.BookRepository
-import com.example.bookie.utils.MyAdapter
+import com.example.bookie.utils.FeedItemsAdapter
 import com.example.bookie.utils.OnScrollListener
 import com.example.bookie.utils.SnackbarUtil
 import com.github.salomonbrys.kodein.KodeinInjector
@@ -59,7 +58,7 @@ class BookList : AppCompatActivity() {
     private fun buildList(books: MutableList<BookFeed>, query: String) {
         val recList = findViewById<RecyclerView>(R.id.feed_container)
         val viewManager = LinearLayoutManager(applicationContext)
-        val viewAdapter = MyAdapter(books, this)
+        val viewAdapter = FeedItemsAdapter(books, this)
         viewManager.orientation = LinearLayoutManager.VERTICAL
 
         recList.apply {
